Fix `tpmsimulator` tags for tests
