+++ tpm/rand/rand_simulator_test.go
@@ -1,3 +1,6 @@
+//go:build tpmsimulator
+// +build tpmsimulator
+
 package rand
 
 import (
